Fix OAuth token exchange crash for perpetual license plans
The server returns expires_in: 3153600000 (36500 days * 86400 sec/day)
for perpetual plans, which overflows Int32.MaxValue (2,147,483,647).
System.Text.Json throws FormatException when deserializing the value
into the int ExpiresIn property on OAuthTokenResponse.

Changed ExpiresIn from int to long on the C# side.

+++ desktop/PrivStack.Services/PrivStackApiClient.cs
@@ -473,7 +473,7 @@ public record OAuthTokenResponse
     public string TokenType { get; init; } = string.Empty;
 
     [JsonPropertyName("expires_in")]
-    public int ExpiresIn { get; init; }
+    public long ExpiresIn { get; init; }
 
     [JsonPropertyName("cloud_config")]
     public OAuthCloudConfig? CloudConfig { get; init; }
